pyproject.toml
[project]
name = "msrating"
version = "0.1.0"
description = "Веб приложение на Flask, для подсчёта рейтинга внутри клуба MafiaStyle"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "mysqlclient>=2.2.7",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.1",
]

[tool.ruff]
target-version = "py312"

lint.select = ["E", "F", "I", "C", "W"]

lint.ignore = []

line-length = 120

lint.extend-ignore = ["C901", "F403"]
